Minor folder restructuring.
