+++ extensions/amp-ad-network-base/0.1/amp-ad-utils.js
