+++ extensions/amp-ad-template-common/0.1/config.js
