+++ gulpfile.js
@@ -72,14 +72,12 @@ declareExtension('amp-accordion', '0.1', {hasCss: false});
 declareExtension('amp-ad', '0.1', {hasCss: true});
 declareExtension('amp-ad-network-adsense-impl', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-adzerk-impl', 0.1, {hasCss: false});
-declareExtension('amp-ad-network-base', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-doubleclick-impl', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-fake-impl', 0.1, {hasCss: false});
-declareExtension('amp-ad-network-template-base', 0.1, {hasCss: false});
-declareExtension('amp-ad-network-template-common', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-triplelift-impl', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-cloudflare-impl', 0.1, {hasCss: false});
 declareExtension('amp-ad-network-gmossp-impl', 0.1, {hasCss: false});
+declareExtension('amp-ad-template-common', 0.1, {hasCss: false});
 declareExtension('amp-ad-exit', 0.1, {hasCss: false});
 declareExtension('amp-analytics', '0.1', {hasCss: false});
 declareExtension('amp-anim', '0.1', {hasCss: false});
